Use Spark JTS when generating footprints
Also cleans things up by removing the traits that were intended to be
generally useful as part of a failed experiment.

Fixes #113
